Generated by Spark: the test connection button is not registering clicks, the debug test button is clicking but nothing is happening after
the logs are just the time repeated over and over not per line - but just the time again and again and again and again - nothing else

clean this up and focus on making a simple test that conforms to a normal test.

connect, login, handshake, all ok, disconnect.

this would be a good test,

this should be like 40 lines of code not 5 classes and 34 test buttons - a really simple straight forward thing, why is this so complicated for you?

I could write a garbage working version of this