Generated by Spark: https://developers.eveonline.com/docs/services/sso/
add a fully working sso login component - this login should be a seperate component from the manual or admin login.

if a manual login occurs - then a eve esi sso login happens after that - clear the manual login and replace i with the sso login.

lets add a roles system that can be applied to both the manual and sso login users