Generated by Spark: there should be one static in the header login next to the local sign in button with eve online button component : https://web.ccpgamescdn.com/eveonlineassets/developers/eve-sso-login-black-large.png
this makes the two seperate login options always visible in the header.

when logged in with sso - add the character image and corp image to the upper left of the header left of the existing lmeve text

validate how your handling the esi with how its handled at :
https://github.com/roxlukas/lmeve

this lmeve install is our base reference for what were making cause its what i used to use and it works i think mostly but you can fix the bits that dont.