Generated by Spark: asince we have the database scema manager there already with the database schema to send to the database - that should be what is used to create the schema file to send - an option for use sites existing schema defaults or custom file should exist in the schema step and if they use defaults - create and send dynamically.
there is a eve SDE tab - it keeps track of the latest SDE and if were up to date or not - downloading the latest should be a default option - with an option to override with a custom SDE file.

there is a possibility that the database is on localhost or another machine - there should be an indicator of having a proper connection to the database machine with ssh, or if its localhost also green light.
a button should exist to connect with ssh, copy the files over to the database or if the db exists on the localhost handle appropriately (just run them without copying?).

basically the only step the user shouldnt have automated is if its another machine - they would have to validate the ssh inital connection manually. after that all else seems like it can be managed automatically if its in this order...