Generated by Spark: move the database schema and eve sde export blocks should be set side by side merged into the database setup operations as selections for part of the process, not a seperate thing...  so  unift all this into one neat block - currently this is also duplicated mostly kinds -
i can still see the remote database connection below the schema manager - that shoulda been removed on the last step?

change the setup requirements block - remove it and instead move "setup requirements" text (with the question mark icon) to the top, and when its clicked on show a popup containing the basic steps the user should carry out.