Generated by Spark: So in the manufacturing tab lets have two subtabs, one for administration and generating / assigning tasks, and one for viewing active manufacturing jobs, with a filter for jobs assigned to the current user, and all jobs activity in the corp (all assigned and in progress).
Assigning jobs should happen with two drop downs, one you type the item name in to select the item (type in with name filter autocomplete in a drop under to select the right one from.  So as you type the possible list gets shorter due to matching text. Once that's selected the icon should appear and be click able for a pop-up info window from the item. The other typeable drop down is for selecting the pilot to assign it to. Then a text box to select an amount to make and then an assign button,

Jobs do not have rewards in isk, there is an hour worked conversion point system by job type. check the lmeve github code for reference on this element as it requires its own page to display and manage.