Generated by Spark: i cannot edit the bonus values in the administration tab of manufacturing
the upper half of the administration tab could be a manufacturing overview - and then the lower half would be assigning new tasks and below that would be editing pay per hour per job type and multiplier value adjustments.

the bonuses should be labeled :
RUSH / Special Delivery / Excess Work

cant really determine checkbox dropdown and input fields from most of the rest of the documents - makes it easy to not see elements that you could interact with - we need these to be white  or grey then white when you type in them - and the current color when its read only

i saw one of the entries on the job type was unassigned - you could add an action for a member to claim unassigned jobs - these jobs should go on thier own tab so they definately stand out.

job types are missing copying and reactions.

when assigning a job - a station needs to be assigned where the work is intended to be done. the eve blueprint data will provide a materials cost - this should be displayed - and a job should have the option to be auto reoccuring assigned to the same pilot upon completion immediate.....  or upon completion then renewed at the first of the month / week