Generated by Spark: under settings -> corporation - it says database settings configure database core settings - that snot right
there should be a method to do a corp level esi authentication against the logged in character - this i assume is the best way to get the corp esi scope data from someone who just logged in with esi.

a person logs in and you request esi scope data for them the minimum really, enough to get around the site basic - if they tie in to a database username - for being assigned jobs etc they need to approve a larger scope of esi stuff for the additional functionality

if they are a ceo or directior for a corporation in eve they can auth the corp scope esi data.

if we can align login with esi and tie the logic in for these things then i think we will have enough esi functionality that i can actually test it with my characters