Generated by Spark: it does not seem like the setup ssh connection button worked - nothing happened when i clicked it but deploy scripts lighting up what should happen : if there is a valid host ip (any ip) setup ssh connection should run the script to attempt to connect to the ip via ssh - debug data should appear in the conneciton logs - rrelevant connection strings should be forwarded to the logs .. connecting, authenticating, awaiting user approval on remote machine.... recieved, ssh connected or timed out - and set the status accordingly. this being a go would enable deploy scripts - which should create a folder on the remote machine and process teh scripts to handle their intended jobs. once the scripts come back as yay done! then run remote setup lights up. if the ssh status is not green, then control pad should also say : "Requires Config" - add another questionmark icon here and a popup providing the explination of what the 3 buttons do and that they need to initally manually approve the ssh request on the remote machine
test connection looks for the lmeve user on testing - if the remote setup hasnt been setup yet its very possible the machine exists and if connected to sql then its a good connection ready to work with a database on the other end.  looking for the lmeve user would imply that yes there is a sql connection, no its not setup yet, the logs should reflect this. mention ok on connect, but maybe ask if the database and or the username / password has been configured properly.  knowing the actual fail along these setup steps will go a long way.

since were focusing on the settings page lets get the lgoic for the check sde button working too.  it should check for the release date of the latest fuzzwork content and compare it to the last known update or the inital install (time only)

lets add a button to the control pad that updates the SDE - this should only be interactable if all else is online and working well