Generated by Spark: test connection looks for the lmeve user on testing - if the remote setup hasnt been setup yet its very possible the machine exists and if connected to sql then its a good connection ready to work with a database on the other end. looking for the lmeve user would imply that yes there is a sql connection, no its not setup yet, the logs should reflect this. mention ok on connect, but maybe ask if the database and or the username / password has been configured properly. knowing the actual fail along these setup steps will go a long way.
since were focusing on the settings page lets get the logic for the check sde button working too. it should check for the release date of the latest fuzzwork content and compare it to the last known update or the inital install (time only)
the check sde should get the proper latest release date from the proper downloadable from https://www.fuzzwork.co.uk/dump/

lets add a button to the control pad that updates the SDE - this should only be interactable if all else is online and working well. theres already functionality to install the sde with the side script package  so we can probably leverage that
lets add a status offline online indicator for the eve server itself in the system status.
and a status indicatior for is there a corporation with active esi scopes ( someone logged in and validated as a ceo / director providing the larger corp esi scopes)