Generated by Spark: test connection should test the ssh port to see if its open for connections on the remote machine
setup ssh connection button still has dummy logic - thats not going to do me any good.

read the scripts folder scripts - we created special scripts to deploy to the remote machine we ssh connected to - thats why we connected to it - the functionality we need is in the scripts were supposed to deploy and run through ssh.
we even scripted make the ssh connection i believe and were supposed to run it from the buttons - and the saftey check is on ssh new key user approval on first connection - so the user will need to be at the other computer to validate that and then the ssh connection completes and we can deploy the scripts - also in the scripts - then we can do the remote setup