Generated by Spark: lets help the scripts a little - the users will have root / sudo control over the machine this is on - so the scripts can and should easily connect to another machine with ssh - no keys need to be generated they are approved with a yes no on the remote machine - the current setup ssh commection does dummy logic - it should : open the connection to the remote server for ssh if it connected good, if not stop if connected, in the bash scripts - await the validation text from the other side, then on the remote machine create a working folder, upload the scripts,
once complete then deploy scripts is done.
at this point the run remote setup which runs the freshly moved scripts on the remote machine.
wait for the scripts to complete -
these configure the elements that the browser connection cannot do - read the bash scripts... this is how we get around the limitations. think deeply here.